Web Framework (Wix Clone - Drag and Drop System)
I'm currently developing a web development platform with drag-and-drop functionality akin to that of Wix and GrapeJS.

This framework currently supports the addition of text areas that users can position as desired through a drag-and-drop interface. Additionally, it allows for seamless integration of images within text sections or directly onto the canvas.

Upcoming features include the integration of animation effects utilizing animate.css, alongside the implementation of a responsive design system through CSS Grids complemented by bespoke media queries.